+++ API/DTOs/Bet/Bet/CreateBetDTO.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace API.DTOs;
 public class CreateBetDTO
 {
-    public float value { get; set; }
+    public double value { get; set; }
     public int coinID { get; set; }
     
     [RegularExpression("1|X|2", ErrorMessage = "The type must be either '1', '2' or 'X' only.")]
