+++ API/Interfaces/Wallet/IWalletRepository.cs
@@ -37,7 +37,7 @@ public interface IWalletRepository
 
     #region FUNCTIONS
     Task<bool> CheckCoinExistById(int id);
-    Task<bool> CheckBalance(int appUserId, int coinId, float balance);
+    Task<bool> CheckBalance(int appUserId, int coinId, double balance);
     Task<WalletCoin> GetWalletCoinDetailedAsync(int appUserId, int coinId);
     Task<Coin> GetCoinByIdAsync(int coinId);
