+++ API/Model/Bet/Bet.cs
@@ -4,7 +4,7 @@ namespace API.Model;
 public class Bet
 {
     public int Id { get; set; }
-    public float value { get; set; }
+    public double value { get; set; }
     public int coinID { get; set; }
 
     [RegularExpression("1|X|2", ErrorMessage = "The type must be either '1', '2' or 'X' only.")]
